requirements.txt
slack-sdk
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
certifi